Python App
Functions folder contains different methods for the GA selection, crossover, mutation, and fitness functions.

The Big test python file is where such functions are applied and where users can experiment with changing parameter values.

The read and write python files are used for easier extraction and uniform of the results.